replace share as image implementation
